Consolidate test and some cleanup (#287)
* WIP Sorting and grouping tests. 

* Deprecate some PPE functions

* Skeleton and Summary mostly consolidated

* update return type and bugfix in mergeVariableSolverData!

* test better types, getFactors tags
